Enum plaid::api::client::response::Response

pub enum Response<P: Product> {
    MFA(User, Challenge),
    ProductNotEnabled(User, P),
    ProductData(P::Data),
    Authenticated(User, P::Data),
    Unknown,
}

Response

Represents the response from the last API request. This does not encapsulate any errors, rather it indicates different stages of the user lifecycle.

It does not implement Decodable because it is simply an enum that represents a specific meta data on top of any Decodable types that it wraps.

For error handling, Result is used alongside plaid::api::error::Error.

Variants

MFA

Waiting on multifactor authentication code from the user

ProductNotEnabled

Returned when a request is made for a Product that is not currently enabled for the given User.

If this occurs, you should upgrade the User so that they have access to the Product.

ProductData

We have sucessfully fetched the available data pertaining to the given Product.

Authenticated

We have successfully authenticated the user, and have retrieved the relevant Product::Data along with that authentication.

Unknown

Nothing is known about the user and no requests have been made