Struct plaid::api::client::Client

pub struct Client<'a> {
    pub endpoint: &'a str,
    pub client_id: &'a str,
    pub secret: &'a str,
    pub hyper: &'a Client,
}

Client

Represents a Plaid API consumer. Encapsulates the endpoint, client_id and secret of the consumer.

This is where all requests to the API start.

Fields

endpoint

E.g https://api.plaid.com.

client_id

Your application's client_id.

secret

Your application's secret.

hyper

The instance of hyper::Client to use. In most cases you simply need hyper::Client::new(). However this is a good place to configure things like proxies, timeouts etc.

Methods

impl<'a> Client<'a>

fn request<P: Product>(&self, product: P, payload: Payload) -> Result<Response<P>, Error>

Make a request to the given Product, using a Payload describing the intention of the operation.