Struct plaid::api::data::account::Account

pub struct Account {
    pub id: UID,
    pub item_id: UID,
    pub current_balance: Amount,
    pub available_balance: Option<Amount>,
    pub institution: Institution,
    pub account_type: String,
    pub account_subtype: Option<String>,
    pub account_number: Option<String>,
    pub routing_number: Option<String>,
    pub wire_routing_number: Option<String>,
}

Account

Represents one account associated with the given User.

Fields

id

The unique id of the account.

item_id

An id unique to the accounts of a particular access token

current_balance

The total amount of funds in the account

available_balance

The Current Balance less any outstanding holds or debits that have not yet posted to the account. May sometimes not be available.

institution

The financial institution associated with the account.

account_type

The classification of this account. See here for more info.

account_subtype

A more detailed classification of the account. This is not always available. See here for a list of possible types.

account_number

The user's bank account number. Only available when using api::product::Auth.

routing_number

The user's routing number. Only available when using api::product::Auth.

wire_routing_number

The user's wire routing number. Only available when using api::product::Auth.