Struct plaid::api::data::transaction::Transaction

pub struct Transaction {
    pub id: UID,
    pub account_id: UID,
    pub amount: Amount,
    pub category_id: CategoryID,
    pub context: Context,
    pub categories: Vec<String>,
    pub pending: bool,
    pub date: Date,
}

Transaction

Represents a single transaction associated with a given Account.

Fields

id

The unique identifier of this transaction.

account_id

The associated Account.

amount

Dollar value as as float. It is positive to indicate money moving out of the account, and negative to indicate that money is moving in.

category_id

The category to which this account belongs. A list can be found here.

context

The context in which the transaction occurred.

categories

An hierarchical list of the categories in which this transaction belongs to.

pending

When true, then this transaction is cleared and immutable. When false, then it is posted and subject to change in the future.

date

The date on which the transaction took place. Plaid standardizes using the ISO 8601 format.