Module plaid::api::data::types

Type aliases that map to common types.

Type Definitions

AccessToken
A user's secret access token

Amount
All amounts are represented in a 64-bit floating-point type. This is for legacy reasons and may change in the future.

CategoryID
Category identifiers are represented by an unsigned integer.

ClientID
Your client id from the Plaid dashboard.

ClientSecret
Your client secret from the Plaid dashboard.

Date
Dates are simply stored as their original String representation. It is up to you to parse it with your favorite date/time library.

Institution
A user's institution. See here for a list.

MFACode
A user's multi-factor authentication code.

Name
A user's real namee.

PIN
A PIN number

Password
A user's bank account password.

UID
Unique identifiers in Plaid are represented as a globally unique hash.

Username
A user's bank account username.