Enum plaid::api::mfa::Challenge

pub enum Challenge {
    Code,
    DeviceList(Vec<(Device, String)>),
    Questions(Vec<String>),
    Selections(Vec<(String, Vec<String>)>),
}

Represents one of the different types of multi-factor-authentication challenges Plaid supports.

Todo: support all mfa challenges

Variants

Code

A token-based authorization, this token will be sent to one of the user's registered devices.

DeviceList

A list of possible challenge devices, in which the user should choose one and then pass along the selection using api::client::payload::AuthenticateOptions. It is in the form of (device_type, device_mask).

Questions

A list of questions that need to be answered.

Selections

A list of multi-choice selections