Module plaid::api::product

Product definitions that provide endpoint and response deserialization information.

Reexports

pub use self::connect::*;

pub use self::auth::*;

pub use self::balance::*;

pub use self::info::*;

Modules

auth
Auth is a product that allows you to authorize ACH transaction from the end-users account. It will return account data including account numbers and routing numbers if authorization is successful.

balance
Balance is a product that allows users to query the account balance of a given user.

connect
Connect is a product that Plaid offers. It allows you to retrieve account balance and transaction history data.

info
The Info endpoint allows you to retrieve various account holder information on file with the financial institution, including names, emails, phone numbers, and addresses.

Traits

Product
Anything that implements Product can be used as a product.