Module plaid::api::product::balance

Balance is a product that allows users to query the account balance of a given user.

Endpoint example

use plaid::api::client::{ Client, Response, Payload };
use plaid::api::product;
use plaid::api::types::*;
use plaid::api::user::{ User };

let client = Client { endpoint:  "https://tartan.plaid.com",
                      client_id: "testclient",
                      secret:    "testsecret",
                      hyper:     &hyper };

let user = User { access_token: "testaccesstoken".to_string() };

let response = client.request(
  product::Balance,
  Payload::FetchData(client, user, None))
  .unwrap();

Structs

Balance
Balance is the endpoint you need to fetch transaction for a User

BalanceData
Representation of data that is retrieved from the Balance product.