Module plaid::api::product::income

The Income endpoint allows you to retrieve various information pertaining to a user's income.

Endpoint example

use plaid::api::client::{ Client, Response, Payload };
use plaid::api::product;
use plaid::api::types::*;
use plaid::api::user::{ User };

let client = Client { endpoint:  "https://tartan.plaid.com",
                      client_id: "testclient",
                      secret:    "testsecret",
                      hyper:     &hyper };

let user = User { access_token: "testaccesstoken".to_string() };

let response = client.request(
  product::Income,
  Payload::FetchData(client, user, None))
  .unwrap();

Structs

Income
The Income endpoint.

IncomeData
The underlying data representation of Income.

IncomeInternalData
Internal data representation of the income response

IncomeStream
An income stream represents a stream of income that Plaid has detected from their transactions.