Struct plaid::api::product::income::IncomeInternalData

pub struct IncomeInternalData {
    pub income_streams: Vec<IncomeStream>,
    pub last_year_income: Amount,
    pub last_year_income_before_tax: Amount,
    pub projected_yearly_income: Amount,
    pub projected_yearly_income_before_tax: Amount,
    pub max_number_of_overlapping_income_streams: u64,
    pub number_of_income_streams: u64,
}

Internal data representation of the income response

Fields

income_streams

A list of income streams.

last_year_income

The sum of user's income over the past 365 days. If we have less than 365 days of data this will be less than a full year's income.

last_year_income_before_tax

last_year_income interpolated to value before taxes. This is the minimum pre-tax salary that assumes a filing status of single with zero dependents.

projected_yearly_income

User's income extrapolated over a year based on current, active income streams. Income streams become inactive if they have not recurred for more than two cycles. For example, if a weekly paycheck hasn't been seen for the past two weeks, it is no longer active.

projected_yearly_income_before_tax

projected_yearly_income interpolated to value before taxes. This is the minimum pre-tax salary that assumes a filing status of single with zero dependents.

max_number_of_overlapping_income_streams

Max number of income streams present at the same time over the past 365 days.

number_of_income_streams

Total number of distinct income streams received over the past 365 days.